Made couple more edits
I noticed you added the cloud distribution section and was wondering if we want to talk a little about the benefits of a distributed configuration? Like the users always have the most up to date config since it's cached every time... I tried to imply this. Let me know what you think.
For the links in the advanced topics section... The community link is an external link, so it's visible in the text. The style guide states to use doc name or a term reference in the support guide as a link. See Style Guide... https://wiki.autodesk.com/pages/viewpage.action?spaceKey=SHOT&title=Style+Guide#StyleGuide-Linkingtothesupportsite  I left the tank links in, but moved the link to only the words "here"...