Adding tank cache_app instructions
Quite a few users will use this guide to figure out how to add apps that have never been added before, as such we should inform them of the `tank cache_apps` required step if they are using a centralized configuration.
Also modified the way the Linux/Mac vs Windows tank commands are listed. Changed the push command so that the windows version is not a Note.